ref(Architecture): BookService 계층 구조 분리
