+++ src/main/java/com/Familyship/checkkuleogi/domains/book/service/BookServiceImpl.java
@@ -1,51 +1,25 @@
 package com.Familyship.checkkuleogi.domains.book.service;
 
-import com.Familyship.checkkuleogi.domains.book.dto.BookCachingItem;
-import com.Familyship.checkkuleogi.domains.book.dto.request.BookLikeRequest;
-import com.Familyship.checkkuleogi.domains.book.dto.request.BookUpdateRequest;
+import com.Familyship.checkkuleogi.domains.book.domain.Book;
 import com.Familyship.checkkuleogi.domains.book.dto.response.BookResponse;
+import com.Familyship.checkkuleogi.domains.book.implementation.BookAdminManager;
 import com.Familyship.checkkuleogi.domains.book.implementation.BookManager;
 import com.Familyship.checkkuleogi.domains.book.implementation.mapper.BookDtoMapper;
-import com.Familyship.checkkuleogi.domains.book.implementation.BookAdminManager;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
-import com.Familyship.checkkuleogi.domains.book.domain.Book;
-import com.Familyship.checkkuleogi.domains.book.dto.request.BookMBTIRequest;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
-
 @Service
 @AllArgsConstructor
-public class BookServiceImpl implements BookService {
+public abstract class BookServiceImpl implements BookService {
 
     private final BookDtoMapper bookDtoMapper;
     private final BookManager bookManager;
     private final BookAdminManager bookAdminManager;
 
-    @Override
-    @Transactional
-    public BookResponse createBook(BookMBTIRequest req) {
-        Book book = bookAdminManager.createBook(req);
-        return bookDtoMapper.toBookResp(book);
-    }
-
-    @Override
-    @Transactional
-    public void deleteBookById(Long bookId) {
-        bookAdminManager.deleteBook(bookId);
-    }
-
-    @Override
-    @Transactional
-    public BookResponse updateBook(Long bookId, BookUpdateRequest request) {
-        Book updatedBook = bookAdminManager.updateBook(bookId, request);
-        return bookDtoMapper.toBookResp(updatedBook);
-    }
-
-    @Override
     @Transactional(readOnly = true)
     public List<BookResponse> getAllBooks() {
         List<Book> books = bookAdminManager.getAllBooks();
@@ -54,25 +28,9 @@ public class BookServiceImpl implements BookService {
                 .collect(Collectors.toList());
     }
 
-    @Override
     @Transactional(readOnly = true)
     public BookResponse selectBookBy(Long childIdx, Long bookIdx) {
         Book book = bookManager.selectBookBy(childIdx, bookIdx);
         return bookDtoMapper.toBookResp(book);
-    }
-
-    @Override
-    public void feedbackOnBook(BookLikeRequest req) {
-        bookManager.feedbackOnBook(req);
-    }
-
-    @Override
-    public List<BookCachingItem> getRecentlyViewedBooks(Long childIdx) {
-        return bookManager.getRecentlyViewedBooks(childIdx);
-    }
-
-    @Override
-    public List<BookCachingItem> getLikedBooks(Long childIdx) {
-        return List.of();
     }
 }
